refactor(backend): move models around
