+++ crates/models/common/src/lib.rs
@@ -7,7 +7,7 @@ use enum_meta::{meta, Meta};
 use enums::{EntityLot, ExerciseEquipment, ExerciseMuscle, MediaLot};
 use rust_decimal::Decimal;
 use schematic::{ConfigEnum, Schematic};
-use sea_orm::{prelude::DateTimeUtc, FromJsonQueryResult};
+use sea_orm::{prelude::DateTimeUtc, FromJsonQueryResult, FromQueryResult};
 use serde::{Deserialize, Serialize};
 use serde_with::skip_serializing_none;
 use strum::{Display, EnumIter};
@@ -227,6 +227,67 @@ pub struct DailyUserActivityHourRecord {
 pub struct DateRangeInput {
     pub end_date: Option<NaiveDate>,
     pub start_date: Option<NaiveDate>,
+}
+
+#[derive(Debug, Serialize, Deserialize, Enum, Clone, Copy, Eq, PartialEq, Display)]
+#[strum(serialize_all = "snake_case")]
+pub enum DailyUserActivitiesResponseGroupedBy {
+    Day,
+    Month,
+    Year,
+    Millennium,
+}
+
+#[derive(Debug, Default, Serialize, Deserialize, InputObject, Clone)]
+pub struct DailyUserActivitiesInput {
+    pub date_range: DateRangeInput,
+    pub group_by: Option<DailyUserActivitiesResponseGroupedBy>,
+}
+
+#[derive(Debug, Default, SimpleObject, Serialize, Deserialize, Clone, FromQueryResult)]
+pub struct DailyUserActivityItem {
+    pub day: NaiveDate,
+    pub total_metadata_review_count: i64,
+    pub total_collection_review_count: i64,
+    pub total_metadata_group_review_count: i64,
+    pub total_person_review_count: i64,
+    pub user_measurement_count: i64,
+    pub workout_count: i64,
+    pub total_workout_duration: i64,
+    pub audio_book_count: i64,
+    pub total_audio_book_duration: i64,
+    pub anime_count: i64,
+    pub book_count: i64,
+    pub total_book_pages: i64,
+    pub podcast_count: i64,
+    pub total_podcast_duration: i64,
+    pub manga_count: i64,
+    pub movie_count: i64,
+    pub total_movie_duration: i64,
+    pub show_count: i64,
+    pub total_show_duration: i64,
+    pub video_game_count: i64,
+    pub total_video_game_duration: i64,
+    pub visual_novel_count: i64,
+    pub total_visual_novel_duration: i64,
+    pub total_workout_personal_bests: i64,
+    pub total_workout_weight: i64,
+    pub total_workout_reps: i64,
+    pub total_workout_distance: i64,
+    pub total_workout_rest_time: i64,
+    pub total_metadata_count: i64,
+    pub total_review_count: i64,
+    pub total_count: i64,
+    pub total_duration: i64,
+}
+
+#[derive(Debug, Serialize, Deserialize, SimpleObject, Clone)]
+pub struct DailyUserActivitiesResponse {
+    pub total_count: i64,
+    pub item_count: usize,
+    pub total_duration: i64,
+    pub items: Vec<DailyUserActivityItem>,
+    pub grouped_by: DailyUserActivitiesResponseGroupedBy,
 }
 
 #[derive(
